[feat] now watch game is avaliable
