+++ src/molecule/GamePlayerInfoBar.tsx
@@ -25,7 +25,7 @@ const GamePlayerInfoBar = () => {
 		}
 		return (
 			<div className="flex my-6 ml-6">
-				<img className={`shrink-0 w-14 h-14 rounded-full`}
+				<img className={`shrink-0 w-14 h-14 rounded-full shrink-0`}
 						 src={lPlayerProfileImg ? lPlayerProfileImg : "src/assets/default_profile_image.png"}
 						 alt={"user profile"}
 				/>
@@ -47,7 +47,7 @@ const GamePlayerInfoBar = () => {
 					<span>{playerInfo.rPlayerInfo.nickname}</span>
 					<span>{playerInfo.rPlayerInfo.mmr}</span>
 				</div>
-				<img className={`shrink-0 w-14 h-14 rounded-full`}
+				<img className={`shrink-0 w-14 h-14 rounded-full shrink-0`}
 						 src={rPlayerProfileImg ? rPlayerProfileImg : "src/assets/default_profile_image.png"}
 						 alt={"user profile"}
 				/>
@@ -63,26 +63,6 @@ const GamePlayerInfoBar = () => {
 		)
 	}
 
-	const LeaveWatchingButton = () => {
-		if (gameStatus === "WATCHING") {
-			return (
-				<div className="flex flex-row bg-neutral-800">
-					<Spacer />
-					<Button className="bg-red-400 text-xs m-2"
-									onClick={() => {
-										setGameStatus("INTRO");
-									}}
-					>
-						LEAVE
-					</Button>
-					<Spacer />
-				</div>
-			)
-		} else {
-			return null
-		}
-	}
-
   /* ========================================================================== */
 
 	/* render =================================================================== */
@@ -96,7 +76,6 @@ const GamePlayerInfoBar = () => {
 				<Spacer />
 				<RightPlayerInfo />
 			</div>
-			<LeaveWatchingButton />
 			<div className="bg-white w-full h-4"></div>
 		</div>
 	)
